requirements.txt
django
django-rest-framework
numpy==1.24.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
langchain==0.1.0
langchain-community==0.0.10
ollama==0.1.7